animation instance / placeholder character / camera control
created an animation instance class and made an animation bp derived from it. the Speed and bIsInAir variables are set in the c++ class and are exposed to the anim bp. to be able to create the anim bp, i needed a skeleton so i imported a placeholder character but didn't import any animations since we don't need them. also elaborated a lil bit on the camera and spring arm. it's now a third person camera but can easily be turned into first person once we have the proper mesh and animations