finished the first, rough version of the item inspection mechanic. needs refinement when actual assets and environment are in the game,
made the object scale up/down to roughly 40% scale on all axis fornow. made it like this cuz i couldn't figure out how to make the object scale according to the viewport size. it may not be very important to our game in the end.
can and should be refined again later. however, without any assets (items to inspect) it's hard to judge whether such a feat would be an absolute necessity.