carrying 2 objects is now working
can carry up to 2 objects at the same time, one in hand and one on back; if the item in hand is dropped (by pressing R), the item on the back becomes in hand. when dropped, items just stay in the location when button is pressed. tried to simulate physics to make them drop to the floor but a bug kept happening where the item can't be inspected again after it's been dropped.